# Independent solver get initial values