The previous implementation simply passes the entire constraint forward
without any factoring of the constraint at all. This is a problem
since it is highly likely that there are cached solutions to pieces
of the constraint. The new implementation breaks the entire
constraint down into its requisite factors and passes each piece
forward, one by one, down the solver chain. After an answer is
returned, it is integrated into a larger solution. Since, by
definition, no factor can affect another, we can safely create a
solution to the larger constraint from the answers of its smaller
pieces.

After completion, the changes:
1) Pass all provided tests (other than ones that were expected to fail)
2) Have been verified via -debug-validate-solver by a complete run through of 9 members of the bin-utils for
--search=dfs with a halt being placed, and states dumped, after 500 forks
--search=default with a halt being placed, and the states dumped, after 1000 forks
3) Achieved significant speedup on test programs
--search=dfs
a. On average 10% improvement (for example strings saw a decrease from 252 seconds to 162 seconds)
b. Worst case (objcopy) saw decrease in performance of 27% (950 seconds to 1287 seconds)
--search=default
a. On average 48% improvement (for example regElf saw a decrease from 1078 seconds to 884 seconds)
b. Worse case (objcopy) saw decrease in performance of 15% (862 seconds to 1022 seconds)