@holycrap872 2015-03-07
@ccadar , I was focusing on the other pull requests and I complete forgot to fix the spacing/apply my newfound understanding of c++ to this commit. I'll work on it tomorrow then repush it. All the ideas are there but as you noticed, it's a little rough. Sorry.