@ccadar 2015-03-07
Great, thanks.  While at it, may I recommend not to introduce new terminology with "factors".  Let's keep calling them independent sets (or independent subsets).  I agree that getIndependentConstraints() was not the best name, but let's call it getIndependentSet() rather than getFreshFactor().  Thanks!