@holycrap872 2015-03-11
Sorry for the delay.  I changed it back to getIndependentConstraints plus cleaned up the commits spacing wise.