@holycrap872 2015-04-01
@ccadar , after I pushed the changes you requested, some of you questions got blasted, so I replaced them along with my answers.  

In addition, you noted "The corresponding header file is not included, and TravisCI compilation fails."
This happened after I had pushed the changes we had previously discussed (3/10/15).  I was working on timing information for some other stuff and I suddenly panicked, thinking I had forget (as it seems correctly) to include the timing statement.  I then added this one line and pushed, hoping to get it up quickly so it could be reviewed.  I didn't compile it or make check it, which is obviously not good.  I'll be more careful in the future.  My apologies.