@ccadar 2015-04-01
Thanks for the quick answers, Eric. 

>"can I ask why this check is here? Shouldn't arraysInFactor.size() always be != 0 ?"
> If I recall, in some cases the getAllIndependentConstraintsSets function can sometimes return empty 
> factors. I added this check because, while they are extraneous, they obviously don't affect the overall 
> computation. As for why the function itself can return these empty factors, that I didn't track down. I can 
> get a little more info on it if that would make you feel more comfortable about what's going on.

As far as I can tell, getAllIndependentConstraintsSets() cannot return empty factors (I've just changed your check into an assert(arraysInFactor.size() !=0) and it doesn't get triggered on the regression suite.)  But if it can, could you add a test case exercising this behaviour?


> @ccadar ... you had asked if we should leave the assert statement in the final commit of this pull ?
> request.

> I guess I would say it's completely up to you. I kept it there as proof that the pull request worked. If
> you don't think it's necessary, then I can definitely drop it from the pull request.

I am wondering if that check contributes in any way to the performance degradation you see on some benchmarks.  Is it easy to confirm?