@holycrap872 2015-04-01
The reason for degradation in performance has very little to do with the assert statement.  Instead, the reason for the degradation is that the new computeInitialValues function produces smaller but more numerous sub-queries.  These sub-queries may, in turn, require expensive calls to the UBTree's Super/Subset functions.  Eventually, given enough of these sub-queries, the cost of solving all of the small queries WITHOUT a call to the SMT solver may end up exceeding the cost of a SINGLE call to the SMT solver.  Adding the QuickCache (pull request 206) ends up eliminating these costs.  When only examining this particular pull request, however, it is possible for the performance to degrade.