@holycrap872 2015-04-01
In reference to the assert statement, I reran the bin-utils experiments with the statement "assert(arraysInFactor.size() != 0)"  I ended up getting an assertion violation, so clearly something is going on.  I'll dig a little deeper to see where this is coming from.