@holycrap872 2015-04-01
I was able to find an example of a factor that cause the assertion violation...

(Eq 0
     (And w32 (ZExt w32 (ReadLSB w16 352 const_arr2))
              8192))

This factor comes from a query with 67 expressions.  All the other 35 factors in the query seem to make sense.

I think what is going on here is that for some reason this value wasn't concretized earlier in the solver chain.  It results in an empty factory because const_arr2 technically doesn't involve any Array that should be symbolic.  Would you like me to keep digging to see why this expression isn't concretized, or would you rather just keep 
 
if(arraysInFactor.size() == 0){ 
    continue; 
}