@ccadar 2015-04-01
Thanks, Eric.  I think it would be quite interesting to debug this, perhaps there's a missing constant folding opportunity.  I'll try to take a look at this too.  

At the very least, it would be useful to see if we can come up with a small test case that illustrates this issue.