@holycrap872 2015-04-01
No problem. I added the line

assert(eltsClosure.elements.size() > 0);

to the computeTruth(), computeValidity() and computeValues() functions in IndependentSolver.cpp.   When I ran it on the bin-utils program that generated the program, the assertion in the computeTruth() function caught the ref<Expr> I explained above as soon as it was introduced to the constraint.  It didn't have any false assertion violations up until that point.

When I run the test suite with these two assertions, there are two failures.  The first is Solver/LargeIntegers.pc and the second is Expr/Evaluate.pc.  Both deal only with Kleaver, though.  I honestly don't 100% understand what is expected of these test cases so I don't know whether they would be expected to simplify away the constant array.  Are these acceptable, or is a program that generates these problems preferable?