@ccadar 2015-04-01
A program is certain preferable, but I'll take a look at these tests too.