@ccadar 2015-04-02
I took a look at these tests, and they are not that interesting because they show cases in which there are no symbolic variables.  But it would be good to understand why the expression you mentioned does not get folded early on.  Do you have more info on that?  Maybe try to extract a small example?