@holycrap872 2015-04-02
Will do.  I'm working on some other stuff for school but I'll try to have it done by Monday.