@holycrap872 2015-04-03
I found what I believe to be the basis of the problem.

In the function lib/Expr/Constraints.cpp::rewriteConstraints(ExprVisitor &visitor), an existing query can be simplified to absurdity without being removed.  In particular, consider the example where the ConstraintManager had the constraint

===
(Eq 0
     (And w32 (ZExt w32 (ReadLSB w16 N0:(Extract w32 0 (Add w64 256
                                                                (Mul w64 2
                                                                         (SExt w64 (ZExt w32 (Read w8 2 arg0)))))) const_arr1))
              8192))
===

in it's constraint list.  If the new expression 

===
(Eq 45 (Read w8 2 arg0)) 
===

comes along, then the constraint above is simplified to 
===
(Eq 0
     (And w32 (ZExt w32 (ReadLSB w16 346 const_arr1))
              8192))
===

There is, however, no mechanism in the rewriteConstraints function to remove this now concretized constraint.  My question at this point is whether anything "bad" would happen by simply removing any constraint that doesn't contain a reference to a symbolic array.  My main concern would be somehow this simplified constraint being involved in some future branch statement.  As I previously noted, I've seen constraints containing on "const_arr"'s involved in a query to solver chain.  I am afraid by removing this constraint from the ConstraintManager, there might be a reference to a deleted memory object.  What are you thoughts?